ai-agent/requirements.txt
fastapi
groq
python-dotenv
SQLAlchemy
uvicorn
agno
google-genai